Added files via upload
Hey guys, so I tried to get some time in at school to work on this. Not having a Mac sucks for this course. The master project was bugged when I pulled it, and I couldn't even build it since it was stuck on Generic IOS Device. No idea, spent time trying to figure it out. Anyways, here's a new one with some of the layout made using UICollectionViews. I'll be working on it more on Friday which is the next time I have free time.

And I know this is a zip.